syntax error fixed, update supp
