+++ src/parser/handler/pipeline_sequence_handler.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/02/12 11:14:04 by spyun         #+#    #+#                 */
-/*   Updated: 2025/03/04 10:32:53 by spyun         ########   odam.nl         */
+/*   Updated: 2025/03/05 17:53:28 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,44 +33,6 @@ static t_ast_node	*handle_pipe_left(t_token **token)
 		left = redir;
 	}
 	return (left);
-}
-
-static void	process_redirection(t_ast_node *cmd_node, t_token **current)
-{
-	if (!*current || !(*current)->next || (*current)->next->type != TOKEN_WORD)
-		return ;
-	add_redirection(cmd_node, (*current)->type, (*current)->next->content,
-		(*current)->next->quote_type);
-	*current = (*current)->next->next;
-}
-
-static t_ast_node	*handle_pipe_redirection(t_token **token)
-{
-	t_ast_node	*cmd_node;
-	t_token		*current;
-	int			has_command;
-
-	cmd_node = create_ast_node(TOKEN_WORD);
-	if (!cmd_node)
-		return (NULL);
-	current = *token;
-	has_command = 0;
-	while (current && (is_redirection(current)
-			|| current->type == TOKEN_WORD || current->type == TOKEN_WILDCARD))
-	{
-		if (is_redirection(current))
-			process_redirection(cmd_node, &current);
-		else
-		{
-			has_command = 1;
-			add_arg_to_node(cmd_node, current->content, current->quote_type);
-			current = current->next;
-		}
-	}
-	if (!has_command && !cmd_node->redirections)
-		return (free_ast(cmd_node), NULL);
-	*token = current;
-	return (cmd_node);
 }
 
 static t_ast_node	*handle_pipe_right(t_token **token, t_ast_node *left)
@@ -107,15 +69,15 @@ t_ast_node	*parse_pipe_sequence(t_token **token)
 
 	if (!token || !*token)
 		return (NULL);
-	if ((*token)->type == TOKEN_WORD)
+	if (is_redirection(*token))
 	{
-		left = handle_pipe_left(token);
+		left = handle_pipe_redirection(token);
 		if (!left)
 			return (NULL);
 	}
-	else if (is_redirection(*token))
+	else if ((*token)->type == TOKEN_WORD)
 	{
-		left = handle_pipe_redirection(token);
+		left = handle_pipe_left(token);
 		if (!left)
 			return (NULL);
 	}
