+++ src/parser/handler/redirection_handler.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/20 21:55:07 by spyun         #+#    #+#                 */
-/*   Updated: 2025/03/01 13:37:28 by spyun         ########   odam.nl         */
+/*   Updated: 2025/03/05 17:37:06 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
