+++ src/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/20 13:02:50 by spyun         #+#    #+#                 */
-/*   Updated: 2025/03/05 09:36:50 by spyun         ########   odam.nl         */
+/*   Updated: 2025/03/05 17:48:02 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
